Restructure project
Now there is just a single `src` folder for REST. The application in `src`
covers all usage cases explained in the documentation.

The documentation has been reorganized as well: now the `rest` folder only
contains `.md` documents.

In addition, some of the examples have been migrated to the MongoDB-based
implementation.